Door Stuck Pt3: Shutter Handling (#7771)
# About the pull request

This PR is a follow up to #7749 and #7727 adding another exemption to
what should not block a door closing: shutters. I am assuming any
anchored door in the locs of the door can be deemed a shutter, but need
be we can explicitly check for pod doors.

# Explain why it's good for the game

Fixes #7769 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://youtu.be/5TWtsn3tDX4

</details>


# Changelog
:cl: Drathek
fix: Fixed doors considering shutters as blocking
/:cl: