Adds the CIA, Removes UAAC-TIS (#7757)
# About the pull request

Adds the CIA, the intelligence agency that needs to introduction.
Featuring a briefcase full of goodies, covert operations, and several
disguises, CIA agents can be planted into various places or even into
the USCM.

This PR fully replaces the UAAC-TIS intelligence officers formerly in
the game. Nobody really used them (save for a select few) and I figured
it'd be better to go out with the American Gestapo and in with actual
spies.

Note that this is admin spawn only for now. In the future it's possible
additional covers, ERT or even survivor roles.

The following CIA covers have been added:

- Intelligence Analyst: The classic "tourist" business casual look of
Vietnam-era CIA. Casual clothing, civilian rank, modifiable job.
- USCM Liaison: A (**real**) military officer who works with the CIA as
a military liaison, though not in an outright public way. Can be either
O-2 or O-3.
- Marine Raider: A CIA agent operating as a Marine Raider. They're on
the same side, just possibly with differing objectives.
- UPP Soldier: A UPP grunt recruited as a CIA informant.
- UPP Senior Officer: A UPP _Starshiy Leytenant_ officer recruited as a
CIA informant.
- CLF Engineer: A CLF field technician recruited as a CIA informant.

All of the above come with the CIA briefcase, a device capable of
procuring various goods and weaponry for use in spycraft. The briefcase
includes things like:
- Abduction and sedation weapons like chloroform, a tranq gun, and
handcuffs.
- Basic stationery (pen and paper).
- Marine armor and clothing, for when shit hits the fan.
- Extra magazines.
- Hacking stuff.
- Radio keys for various factions, including the CIA's own radio
channel.
- Listening bugs.
- A USCM codebook for verifying you are actually with the government, in
case anyone has doubts.

# Explain why it's good for the game

Spycraft! I believe adding CIA adds a ton of roleplay potential with
these people as well as the ability to carry out actual
(admin-sanctioned) antaggery like abducting people or interrogating
others in a subtle, government-sanctioned way.

# Things CIA Might Do
Event-runners use this table as your inspiration to find out what
morally-questionable things your agents may do.

- Wiretap dialog in certain important areas like the CIC or brig.
- "Extract" an asset like a person or a thing from the Almayer or AO.
- Perform "enhanced interrogation methods".
- Subtly feed information or misinformation to command to sway them a
little.
- Make a problematic person _disappear_.
- Local whistleblower found dead with 5 shots to the back of the head;
ruled self-inflicted.

# The Elephant

CIA agents, naturally, work for the government of the United Americas,
and nobody else. The USCM and CIA are fighting for the same team, but
CIA might have interests or methods that differ slightly or are above
the paygrades of most marines.

CIA agents prioritize plausible deniability for covers with allies, and
they're often well aware of the fact that people can see right through
their generic cover identities; the goal isn't to trick allies, it's to
prevent the CIA from being finger-pointed at directly if something goes
south. The USCM knows there might be agents in their ranks, but since
they're usually actually helpful - and pointing out a CIA agent does
nothing for them anyway - they don't really care, and neither should
marines. (Spies are a different story.)

For obvious reasons no CIA agent should ever say "I am with the CIA." If
not outright disguised, it's up to the individual (or the admin) to
fabricate a cover story, like "I'm an intelligence analyst working for
ARMAT" or "I'm part of a battlefield survey team to improve logistics."
You can modify your ID accordingly to suit your cover. Be creative, and
be subtle.

# Screenshots
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/debdc096-633a-4821-81a2-df3c99979514)

![image](https://github.com/user-attachments/assets/59120e6a-6e41-4aef-ba78-3c529f1cd78d)

![image](https://github.com/user-attachments/assets/53b8679c-6099-4ea9-b1b7-4a21b323b729)

![image](https://github.com/user-attachments/assets/7b8c119b-6b1f-4b3b-8fe1-e34f5462ce0c)

![image](https://github.com/user-attachments/assets/3d6e92e0-c394-4f9a-a4f5-306cb1d38e2f)

</details>


# Changelog
:cl:
add: Adds CIA agents to the game. These are disguised as either
civilians, liaisons, or people in other factions and have a plethora of
tools to achieve any objectives they might be assigned to.
del: Deletes the UAAC-TIS from the game.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>