Queen cannot forfeit in a tunnel (#7773)
# About the pull request

This PR adds an exception to the xeno count for distress signal where if
no xenos were counted, then the living queen(s) are checked. I also
fixed/refactored check_queen_status (used to check if xenos fail to
evolve to queen in 10 minutes) to properly check hives that would count
towards round.

# Explain why it's good for the game

Fixes #7762

Doesn't make any sense why a queen would cower in a tunnel in hopes of
making the marines go away. But it is likely if the hive is abandoning
the old hive core they could potentially all be in a tunnel and
mistakenly trigger this.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/b2697343-4a2c-4ab7-8e78-6a03f885eaf8)

</details>


# Changelog
:cl:
balance: Queen prevents round end even when in a tunnel
fix: Fixed check_queen_status checking any queen not just queens that
matter to round end
/:cl: