Adds a custom HealthMate HUD to CMO vendor that combines a HealthMate and Reagent HUD (#7392)
# About the pull request

This PR adds custom HealthMate HUDs to the CMO vendor that have the
functionality of both a HealthMate and reagant scanner hud.

# Explain why it's good for the game

CMOs are expected to both keep tabs on research upstairs and medical
downstairs. To do this they effectively, they need to be able to see
chemicals inside containers but also have a healthmate on to do their
medical part of their job. This leads to an awkward switching of huds
that only the CMO does whenever they step into research, so you need to
keep a reagent googles somewhere on you at all times if you dont want to
backtrack to your office every time you forget it on your way to
research.

This both rectifies that mostly quality of life issue and gives CMOs
some cool fluff googles to show off with, adding a tiny bit more to the
role. It needs research skills so the doctors cant use it, but
researchers can.

# Testing Photographs and Procedure

Tested on a up to date private instance of CM.
The sprite of the glasses are the same as healthmate HUD glasses.

# Changelog

:cl: Drdodo
add: Added custom HealthMate HUDs to CMO vendor which combine HealthMate
and Reagent scanner HUDs that requires research skill to use
qol: Removed the old reagent scanner and healthmate HUDs from the CMO
vendor
/:cl:

---------

Co-authored-by: Isaac Alp Kocyigit <alpkoc27@students.cs.ubc.ca>