NERFS scout's engineering skill to NOVICE (#7372)
# About the pull request
Scout now have SKILL_ENGINEER_NOVICE instead of SKILL_ENGINEER_TRAINED.

# Explain why it's good for the game
The only reason for this PR is to fix the situation where it is too easy
for scouts to repair the sensor tower.
 
Scouts used to carry engineers in the tarp in order to repair the sensor
tower which was pretty funny. But then they found out that all they
needed were tools. Anyway it's pretty lame that an invisible entity can
repair the tower with little to no risks. Checking the tower every two
minutes to look if it was sneaky repaired really brings me to pre-perma
nesting times. It's lame. And bringing an engi was way cooler and
soulful, and required some actual teamwork. I don't think scout really
needs engineering skills for anything else, C4 is useable with
SKILL_ENGINEER_NOVICE.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: ihatethisengine
balance: scout's engineering skill is now 1.
/:cl: