Type-19 and PPSH Gun Case + Addition To UPP Vendor (#7701)
# About the pull request

Adds two new gun cases, one for the Type-19 SMG, and one for the PPSH
smg.

Adds the Type-19 gun case, Type-19 magazine ammo box and a scale * 1
Type-19 extended magazine to the UPP Requisition vendors as Restricted
Weaponry items.

The PPSH gun case is not currently added to the game, but I intend to
place a few for the UPP on their map/ships.

# Explain why it's good for the game

Grants UPP cargo more options to give unique gear to UPP personnel, plus
giving personnel more options for loadout items.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: A Type-19 submachinegun gun case has been added to the UPP
requisitions vendor as a restricted weapon item, which can be issued to
UPP soldiers as requested.
add: Addition of Type-19 magazines, extended magazines and ammo boxes to
UPP requisiton vendors.
add: A PPSh-17b submachinegun gun case has been added to the game.
Intended to be spawned as a rare weapon, owing to its (in-universe)
collectors and old firearm enthusiasts value.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>