UPP Pilot Job (#7789)
# About the pull request

Adds the UPP Pilot as a job-role for the Faction Clash gamemode. This
pilot has the same role and skill capabilities as its USCM counterpart.

This role does not currently come with any unique or new sprites.
However, I am hopeful that with the end of the sprite freeze, we will be
able to port the far more expansive PVE UPP sprites which will alleviate
this.

# Explain why it's good for the game

With the expansion of UPP dropships, it has now become viable to add a
UPP pilot, who can be trusted to control a dropship and ferry it between
the FOB and the orbiting warship.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: The UPP Pilot has been added as a job-role for the Faction Clash
gamemode.
/:cl:

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>