Makes Resin Salve only go on CD after successful activations, not attempted ones (#7301)
# About the pull request

#6847 introduced a cooldown to the healer drone's salve ability.
Specifically, it was set to apply on an attempted activation of the
ability, not just on a successful activation of it.

This has made the usage of the ability, and thus healer in general, feel
more clunky. It's not an ability like warrior's lunge where the way the
cooldown is applied has considerable balancing concerns, so this moves
the cooldown to successful activations only.

Also removes the notification for the resin salve cooldown expiring.
It's a 0.5 second cooldown, it doesn't need a notification.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

QoL. Makes the main ability of the healer caste feel less clunky and
overall nicer to use.
Considering that the old version of this ability didn't even have a
cooldown, I don't see this affecting balance in any meaningful way.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

https://streamable.com/wuyf3t

</details>


# Changelog
:cl:
qol: Makes Resin Salve only go on CD after successful activations, not
attempted ones
qol: Removes the chat notification for Resin Salve going off CD (it's a
0.5 second CD)
/:cl: