anti c4 gamemode flag, anti drag flag fix (#7459)
# About the pull request

adds togglable flag that prevents c4s to be planeted on humans when it
is on gamemode. fixes bipasses of anti drag

# Explain why it's good for the game

c4ing bodies of other side SUCKS and is hated by every side, if you want
it for zombies, go ahead I place it on HvH. draging with roller beds and
bodybags is not intended and a bug


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: toggleble flag to prevent c4s on human for gamemodes
fix: anti stripmode flag prevents buckling to roller beds and closing
into bags to move dead or crit bodies of enemy faction
/:cl:

---------

Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>