Adds a version of the CO armor without pauldrons (#7143)
# About the pull request

This is based on the old sprite, and is a stat-identical version of the
normal CO armor, minus the pauldrons.

Nessie is behind this, please refer criticism that way. (Unless it's the
code, that's on me)

# Explain why it's good for the game

More choices for drip baby!


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/dd9a32fe-c8fa-40f3-98a2-d98438e74b21)

Can only pick 1:

![image](https://github.com/user-attachments/assets/5640c65e-1b14-4809-8f4a-09ce823c25cc)

Worn:

![image](https://github.com/user-attachments/assets/9c4557b8-934a-4a28-bba8-877a4ffe7b07)


</details>


# Changelog
:cl:
add: COs have a new variant of their armor to choose from
/:cl:

---------

Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>