UPP MMG (#6915)
# About the pull request

Adds in one feature from this now auto-closed PR:
https://github.com/cmss13-devs/cmss13/pull/5945

Adds the T37 MMG for the UPP, their alternative version to the USCM M2C.
The T37 is a stationary weapon, meaning it can not be moved, it has a
slower rate-of-fire but it can sustain firing for significantly longer
before it will overheat.

# Explain why it's good for the game

Useful features for mapping and events. 

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Adds the T37 Medium Machine Gun to the UPP's arsenal. The T37 is an
alternative variant of the M2C HMG, it is a stationary emplacement (that
can not be redeployed) that has a slower rate of fire compared to the
M2C, but it can fire for a far longer period of time before overheating.
This is a purely defensive weapon, and it must be pre-placed on a map by
a mapper.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: Blundir <100090741+Blundir@users.noreply.github.com>
Co-authored-by: harryob <me@harryob.live>
Co-authored-by: cuberound <122645057+cuberound@users.noreply.github.com>
Co-authored-by: vincibrv <vojtechvincibr@seznam.cz>