Suppress runtime for imaginary friend SEA (#7855)
# About the pull request

This PR just adds a nullcheck to `mob_client` for SEA's new player hud.
In this case a flat icon is being generated for a SEA with no client, so
a HUD is not applicable. There isn't an effect by this change though
that I can tell (appearance would be made fine just before and after).

# Explain why it's good for the game
Less runtimes.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/f1d2a0d1-9b2b-4157-b7d4-36f160d16280)

</details>


# Changelog
:cl: Drathek
fix: Fixed a runtime when generating SEA appearance for an imaginary
friend
/:cl: