Hunting grounds is now exempt from affecting larva queue (#7853)
# About the pull request

This PR is a follow up to #7486 which now flags the hunting grounds with
`AREA_YAUTJA_HUNTING_GROUNDS` and can optionally be tested against in
`should_block_game_interaction` which is now checked as far as the ToD
for larva queue.

# Explain why it's good for the game

Generally ghost activities should not penalize you from rejoining the
main game.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Drathek
balance: Hunting ground deaths no longer counts against larva queue
/:cl: