Yautja marks can be used appropriately (#7902)
# About the pull request

This PR allows Yautja to correctly un-mark dead people, and to mark the
dead as gear carriers.

# Explain why it's good for the game

The inability to unmark dead people has caused issues and was not
intentional.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Yautja can now correctly un-mark dead people.
/:cl: