staff and longaxe now in the elite vendor (#7576)
# About the pull request

the staff and the longaxe weapons are now in the elite yautja vendor

Changes the description of the staff to be more fitting.

Also adds back onmobs for the longaxe. Sprites by Thwomper of course
# Explain why it's good for the game

I thought I had added the weapons to both the normal and elite vendors,
but I guess I gaslit myself or something.



# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/aaf43e6d-ce58-4626-a802-66e52a6d387d)

![image](https://github.com/user-attachments/assets/9da946f8-05d7-423b-b0bb-bfe64d694f15)

![image](https://github.com/user-attachments/assets/b0cc5ae7-0a9d-4fd4-aa79-f14904bae330)

![image](https://github.com/user-attachments/assets/d0f53198-3b2c-4650-80a0-e0027fd673dc)


</details>


# Changelog
:cl: Nomoresolvalou, Thwomper
add: Added staff and longaxe weapons to the elite yautja vendor.
imageadd: Adds onmob sprites for the longaxe.
spellcheck: Changes hunter staff description to be more in line with
other weapons.
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>