CPR Buddy 9000 (#7092)
# About the pull request

Adds a brand new bot type called the CPR bot, which is able to target
and locate downed marines and perform CPR automatically.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

I have tested this CPR bot many times; it showed no problems or bugs
that are known to me at least.
it functions as it should.

# Explain why it's good for the game

I think this would be a huge change to how synthetics operate on the
ground of operations.
It actually replaces a piece of equipment that was left out to collect
dust for many years.
while the cpr bot is only a little bit better by a few seconds.
It is designed to make using the autocomprossor more fun and an
enjoyable experience for synthetic players.
I think it will encourage synthetics to really think about what to pick
for their experimental equipment, which is normally a good thing!

Code was inspected and edited by Sindorman

# Testing Photographs and Procedure
I have tested the logic of the CPRbot many times, from it's pathfinding
to it's CPR actions performed.
It performs CPR every 7 seconds and only restors 4 seconds to the downed
person's CPR timer, which is done this way for balance reasons so it
doesn't outperform the auto compressor as it doesn't have a battery like
the auto compressor, but when tested, it performed the same with a few
extra seconds gained but performed the same none the less.
It's pathfinding isn't the best as it's made with the Astar logic, which
all other bots use ingame, but it's decent enough to do the job!
If a bot can't find a valid path, it stops marking that patient as a
valid patient and stops looking for it to save on memory power.
cprbot also uses an IFF check to make sure it only assists people of
it's faction from the USCM and no one else, much like the motion
detector functions "same logic."
It also checks if the person is human, which is a strict requirement,
and also checks if there is another cprbot nearby by at least 2 tiles;
otherwise, the cprbot will go in and assist if all checks are clear.

<details>


https://github.com/user-attachments/assets/c65709d7-a043-4a29-a6f4-9e5e94ae4172

</details>


# Changelog

:cl:
add: Added CPR Buddy 9000 to the experimental vendor.
/:cl:

---------

Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>