FIxes Hunting Grounds Escape (#7864)
# About the pull request
Makes the escape from the hunting grounds actually work properly instead
of any item working on the scanner

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Fixes my shit code 
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: Joe Lampost
fix: Fixes the escape from the hunting grounds
/:cl:

---------

Co-authored-by: kiVts <48099872+kiVts@users.noreply.github.com>