Pipes under constructed walls maplint (#7931)
# About the pull request

This PR is a lint against pipes under walls (if they aren't resin).

# Explain why it's good for the game
Pipes are a travel method for xenos, so running pipes under walls makes
that path safer for them and harder for marines to deal with. In general
maintenance of pipes (IC) should be accessible.

# Changelog
:cl: Drathek Nanu
add: Added maplint to check whether pipes run under constructed walls
/:cl:

---------

Co-authored-by: Nanu <59782240+Nanu308@users.noreply.github.com>