RPG Ammo Type Loading Fix (#7939)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Fixes being able to load a rocket array into the standard rocket
launcher if you use the buddy reload system. Also adds a little bit of
feedback for if the weapon isn't in the active hand or if it's already
loaded.

# Explain why it's good for the game
it insists upon itself

# Testing Photographs and Procedure
i tried loading the thing and not loading the thing and loading the
thing when it was loaded and it all worked, trust me


# Changelog
:cl:
add: Added some text feedback to the RPG buddy loading system for when
it isn't wielded or if it's already loaded
fix: Fixed the RPG buddy loading system being able to load invalid
magazines into the launcher
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>