Ship Operations Synthetic (#7743)
# About the pull request
This PR adds a ship oriented Synthetic who is playable if there is no
Survivor Synthetic present.

Issue at hand was although Synthetics readily accept RP, they are not
able to easily go out of their way to initiate it due to their
responsibilities outside of 3 phases: Pre-drop, Surgery, Post-drop.

This is a push to better incentivize the availability of Synthetics to
RP. Currently they are burdened with a multitude of responsibilities
compared to other roles which are able to prioritize RP. This
effectively bridges that gap significantly, and encourages RP whilst by
having 1 Synthetic operate more in-line with the bridge officer / ship
maintenance synthetic within the bounds of lore.

BIG Thanks to Drathek for the role authority code!!!

Great suggestion by Beagle!!
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game

Conducive to RP availability and pretty healthy for immersion within
lore. Does not change the overall amount of Synthetics.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.


![image](https://github.com/user-attachments/assets/7ae830f4-0e7f-4ddb-b83d-2d98006f8391)


![image](https://github.com/user-attachments/assets/d1b51c8c-fc1e-4c05-8556-1f57a72fc946)


![image](https://github.com/user-attachments/assets/c61b9f65-0093-428f-94f0-e9b75b813c32)


![image](https://github.com/user-attachments/assets/5cadf7d3-b005-47b6-a17b-70b606514631)




built ok 
</details>


# Changelog
:cl:
add: Adds the Ship Operations Synthetic. This ship-oriented Synthetic
assists the USS Almayer with various tasks and improves the atmosphere
of the vessel, they maintain the entire ship and act as its pilot. This
role is enabled if there is no Survivor Synthetic present. Huge thanks
to Drathek for roleauthority code!
add: Changes role selection color of WJ to be slightly darker to
differentiate from Synthetics.
code: restructured lobby role selection menu to accomadate new role
/:cl:

---------

Co-authored-by: Drulikar <drathek@gmail.com>
Co-authored-by: harryob <me@harryob.live>
Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>