Removes non combat click delay (#7802)
# About the pull request
Basically #2130 but coded differently
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
#2130
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
del: Removed non-combat click delay, interacting with items not on your
person should be significantly smoother
balance: Clicking on the floor (open turfs) with an item will no longer
cause a click delay
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>