Use revparse to determine head for DMM test (#7936)
# About the pull request

This PR is a follow up to #7595 tweaking how the head commit is
determined. This change basically means when ran locally the test will
be incorrect unless you checkout the PR the same way as github checkout
v4 does (e.g. `git fetch upstream
+0abb4ebc6c24144032afa85686a5569ffda29322:refs/remotes/pull/7595/merge`)

Ultimately the issue appears to be that HEAD on the github runner is
already a merge commit, so this incorrectly makes latest master the
ancestor commit rather than what the contributor's ancestor commit would
be resulting in unnecessary lint failures asking them to run map merge
even if they are running with hooks. Only if they were up to date with
latest master would the contributor's local hook and the linter line up.

See
https://git-scm.com/docs/git-rev-parse#Documentation/git-rev-parse.txt-emltngtltpathgtemegem0READMEememREADMEem
as to why I needed to use HEAD^2 (effectively need to select the C
commit in the diagram so that the ancestor commit is determined to be
something earlier than the B commit).

# Explain why it's good for the game

Hopefully more accurate linting and less requirements to run map merge.

# Changelog

No player facing changes.

---------

Co-authored-by: harryob <me@harryob.live>