Change the color of the Squad Leader pointing arrow to the color of the squad (#7933)
# About the pull request

This will change it so the big green arrow SL's get will now change
color to the respective color of the squad. This also works for or any
other squads that have their own equipment color, like UPP squads.
# Explain why it's good for the game

Makes it easier to spot from which squads SL the arrow is coming from.
Also looks cool.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/user-attachments/assets/3251caf6-2cb4-471f-a319-df2f842f389f



Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Squad Leader pointer arrows are now colored based on the color of
the squad.
/:cl: