Demo Specialist Loader Assistant Role (#7887)
# About the pull request

Copies the implementation of the Spotter role to make one for the Demo
spec as well, called the "Loader", it has no real special abilities,
other than a dedicated body to carry rockets and reload them. Their
pamphlet gives them a little more endurance.

# Explain why it's good for the game

Encouraging teamplay amongst the marines will always be good for the
game. Plus, it's good for roleplay.


# Testing Photographs and Procedure

i compiled, ran the server, went pfc, got the stuff, checked that my HUD
and ID updated correctly, and that i could wear the armor


# Changelog
:cl:
add: Added a Loader assistant role to the Demolition specialist
/:cl: