AMR ammo economy nerf (#7874)
# About the pull request
Nerfs the amount of mags you get from the kit box, and the req crates

https://github.com/cmss13-devs/cmss13/pull/5539#pullrequestreview-1865953845
# Explain why it's good for the game
 #5539
Just like the PR above, this aims to cut down on the amount of ammo you
get as the AMR specialist.
This spec gets about 45 shots just for getting the box, and 3 more extra
magazines you can buy from the ammo vendor, totaling up to 72 shots just
for obtaining the spec kit, i think this is absurd for how viable the
rifle currently is

This nerfs that number down to 45 totalling the ammo you buy, so you
have 45 shots first drop.

This sniper is already very good and a bad addition to the game, the
fact that it never requires req supply (if you dont just shoot at
everything) is bad, specs should be costing the operation money if
theyre going to be viable, like the scout and the sadar. AMR is already
very oppressive and i think this is the best way to approach nerfing it
instead of damage changes or otherweise removing it.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
balance: AMR sniper spec magazines reduced to 2 for the box.
balance: requistions amr sniper spec magazine order reduced to 3 mags.
/:cl:

---------

Co-authored-by: InsaneRed <prodexter31@outlook.comm>
Co-authored-by: harryob <me@harryob.live>