Fix lots of random throwing and explosion runtimes (#7861)
# About the pull request

This PR attempts to mitigate various oversights and runtimes that occur
with explosions and throw_atom. They are really getting out of hand.
This was kicked off by a round in watchtower testing that caused the
master controller to crash from recursive throw code.

# Explain why it's good for the game

Should mean aspects like two crushers colliding at each other even is
handled, not to mention aspects that would making throwing an atom
impossible (like trying to throw an atom at an area - why is an area
considered an atom? - why is an area a result in range??). There are
likely performance gains from this, but then again, code not halting
because of runtime could mean more is being done too...

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/57a89248-a47a-4e40-869c-6fb2d59b2450)

</details>


# Changelog
:cl: Drathek
fix: Fixed various runtimes caused by explosions and throw_atom code
/:cl:

Co-authored-by: harryob <me@harryob.live>