Adds some more professional accessories and suits to the synthetic cosmetics vendor (#7930)
# About the pull request

Adds the new Hybrisa EMS uniforms to the synthetic uniform vendor, a
briefcase, and adds a holo badge to the vendor. For the professionals.

# Explain why it's good for the game

More options for professional type synthetics is good.


# Testing Photographs and Procedure

Ran on local


# Changelog

:cl:
add: Added EMS red and green uniform and jumpsuit, and holobadge to
synthetic uniform vendor. Added a briefcase to the uniform vendor.
/:cl:

---------

Co-authored-by: naut <55491249+nauticall@users.noreply.github.com>