Human Resprites & Tweaks - Marines looking awesome (#7521)
# About the pull request
Fixes, tweaks and resprites the body types and muscularity plus the skin
tones for humans in the game. Female torso's are now a thing as well, as
that is brought back. Underwear and Bra's have also been tweaked to fit
new models.

# Explain why it's good for the game
Visually better and tweaked skin tones to look better plus female torso
added.


# Testing Photographs and Procedure

![oomanExamples2](https://github.com/user-attachments/assets/0ac1ae58-e21c-4cfe-9976-af6ef97bc5ca)



# Changelog
:cl: Nanu, Wei (SleepyNecrons, Esslenek), Spartanbobby, HarryOb
imageadd: Fixes, tweaks and resprites the body types and muscularity
plus the skin tones for humans in the game. Female torso's are now a
thing as well, as that is brought back. You can now be a man and have a
female body, and vice versa. Try not to get catfished Marines. Underwear
and Bra's have also been tweaked to fit new models. Sprites created by
Wei.
/:cl:

---------

Co-authored-by: harryob <55142896+harryob@users.noreply.github.com>
Co-authored-by: Drulikar <drathek@gmail.com>