CAS no longer has infinite NVGs roundstart (#7995)
# About the pull request
Fixes half of #5676 
When #4812 was merged, it gave all CAS pilots infinite, roundstart night
vision with no requirements. This fixes that, making it actually check
if the DS matrix is upgraded or not.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Bugs should be fixed. Having researchers run over to CAS pilots with
photosensitive vials to grant NVGs was sovl and gave research something
to do that wasn't AFKing for points.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/user-attachments/assets/3858d1c6-304d-4112-b374-1556a5aaf8aa)

Tested, it works.

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: CAS no longer has infinite night vision, contact research for
matrix upgrades today!
/:cl: